refactor: gallery layout and pagination
