+++ src/layout/css/_media.scss
@@ -1,13 +1,3 @@
-.media-img {
-  border: 5px solid transparent;
-  cursor: -moz-zoom-in;
-}
-
-.media-img:hover {
-  border-color: $secondary;
-  transform: border-color 0.1s;
-}
-
 .media-modal {
   z-index: 9999;
   position: fixed;
